+++ xarray/core/dataset.py
@@ -1959,6 +1959,7 @@ class Dataset(
         region: Mapping[str, slice] | None = None,
         safe_chunks: bool = True,
         storage_options: dict[str, str] | None = None,
+        zarr_version: int | None = None,
     ) -> Delayed:
         ...