+++ doc/whats-new.rst
@@ -22,7 +22,9 @@ v2023.04.0 (unreleased)
 
 New Features
 ~~~~~~~~~~~~
-
+- Added ability to save ``DataArray`` objects directly to Zarr using :py:meth:`~xarray.DataArray.to_zarr`.
+  (:issue:`7692`, :pull:`7693`) .
+  By `Joe Hamman <https://github.com/jhamman>`_.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~